# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OmniChat omnichannel messaging platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Departments
...
  /webhooks/meta:
    get:
      operationId: verifyMetaWebhook
      tags: [webhooks]
      summary: Meta webhook verification (GET challenge)
      parameters:
        - name: hub.mode
          in: query
          schema:
            type: string
        - name: hub.verify_token
          in: query
          schema:
            type: string
        - name: hub.challenge
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Verification successful
    post:
      operationId: receiveMetaWebhook
      tags: [webhooks]
      summary: Receive Meta webhook events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetaWebhookPayload"
      responses:
        "200":
          description: Acknowledged

  # Stats / Dashboard
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, supervisor, agent]
        departmentId:
          type: ["integer", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [email, name, role]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, supervisor, agent]
        departmentId:
          type: ["integer", "null"]
        avatarUrl:
          type: ["string", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, supervisor, agent]
        departmentId:
          type: ["integer", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean

    # Departments
...
    MetaWebhookPayload:
      type: object
      required: [object]
      properties:
        object:
          type: string
        entry:
          type: array
          items:
            type: object

    # Stats
...