# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OmniChat omnichannel messaging platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  avatarUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}