# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OmniChat omnichannel messaging platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  name?: string;
  role?: UserUpdateRole;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  avatarUrl?: string | null;
  isActive?: boolean;
}